// src/types/misc.ts
export type ErrorDetails = {
	type: string
	data: {
		note?: string
		link?: string
	}
}